Module kinetic::commandsUnstable

Available Kinetic commands

Modules

common

Structs

Get
Get's the value and the metadata for the given key

GetKeyRange
Requests a range of keys between two given keys

GetLog
Get's the requested logs

GetNext
Get's the value and the metadata for the given key following the given key

GetPrevious
Get's the value and the metadata for the key before the given key

GetVersion
Get's the version and integrity for the given key

Put
Stores the value asociated with the key

Enums

Delete
Deletes the key