Module kinetic::commands::pinExperimental

Pin based commands

Pin based commands are used for special commands that are valid when the device is locked or to be locked. These are Unlock, Lock and Erase. This commands must be send over a secure connection to protect the confidentiality and integrity.

Structs

Unlock
Unlocks the device

Traits

PinCommand
Trait representing a Kinetic pin based command