Struct kinetic::responses::GetKeyRangeResponseExperimental

pub struct GetKeyRangeResponse {
    pub keys: Vec<Vec<u8>>,
}

A GetKeyRange command result

A GetKeyRange command result contains an ordered list of keys contained between the start key and the end key.

Fields

keys